# lobster-docs/openapi/lobster.yaml
openapi: 3.1.0
info:
  title: Lobster BFF
  version: 0.1.0
  description: >
    Backend for the Lobster dashboard. It wraps the DFNS custody service and a
    few reads. The custody routes are gated on an API token (LOBSTER_API_TOKEN);
    with no token set the gate is open for local dev. The DFNS reads answer 502
    when the upstream call fails or the credentials are missing, and the sign
    path answers 503 until its env is in place.
servers:
  - url: https://lobster-bff.onrender.com
    description: deployed relay
  - url: http://localhost:8787
    description: local dev (hono)
...
security:
  - lobsterToken: []
  - bearerAuth: []
paths:
...
  /allbridge/tokens:
    get:
      tags: [allbridge]
      summary: Tokens Allbridge Core lists for a source chain
      description: >
        Read-only passthrough to the Allbridge Core API. Allbridge is mainnet
        only, so these routes always target mainnet. None of them are
        token-gated: nothing here signs or holds a secret.
      security: []
      parameters:
        - name: chain
          in: query
          required: true
          schema: { type: string, enum: [ETH, ARB] }
      responses:
        '200': { description: token list, each with symbol, decimals, tokenAddress, bridgeAddress, chainSymbol }
        '400': { description: chain must be ETH or ARB }
        '502': { description: Allbridge call failed }
  /allbridge/quote:
    post:
      tags: [allbridge]
      summary: Quote a USDC transfer into Stellar
      description: >
        Returns amountIn, amountOut, etaSeconds (null when the corridor omits a
        transfer time) and gasFeeOptions keyed by payment method. The USDC pool
        on the Stellar side is closed upstream at the moment, so this route
        answers 502 and names the closed corridor as the reason.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceChain, amount]
              properties:
                sourceChain: { type: string, enum: [ETH, ARB] }
                amount:
                  type: string
                  description: positive USDC amount, at most 6 decimals
      responses:
        '200': { description: quote }
        '400': { description: body failed validation }
        '502': { description: Allbridge call failed }
  /allbridge/status:
    get:
      tags: [allbridge]
      summary: Delivery status of a transfer
      description: >
        Delivery on Stellar is automatic; this is how a caller learns the
        transfer landed. txId is the EVM hash the send returned.
      security: []
      parameters:
        - name: chain
          in: query
          required: true
          schema: { type: string, enum: [ETH, ARB] }
        - name: txId
          in: query
          required: true
          schema: { type: string }
      responses:
        '200': { description: transfer status }
        '400': { description: chain must be ETH or ARB, or txId missing }
        '502': { description: Allbridge call failed }
  /allbridge/raw/approve:
    post:
      tags: [allbridge]
      summary: Build an unsigned ERC-20 approve for the bridge
      description: >
        Build only. The caller's wallet or custodial signer signs and broadcasts;
        this service never does.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceChain, amount, owner]
              properties:
                sourceChain: { type: string, enum: [ETH, ARB] }
                amount: { type: string, description: positive USDC amount, at most 6 decimals }
                owner: { type: string, description: EVM address, 0x plus 40 hex }
      responses:
        '200': { description: rawTx }
        '400': { description: body failed validation }
        '502': { description: Allbridge call failed }
  /allbridge/raw/send:
    post:
      tags: [allbridge]
      summary: Build an unsigned bridge send from EVM to Stellar
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sourceChain, amount, fromAddress, toAddress]
              properties:
                sourceChain: { type: string, enum: [ETH, ARB] }
                amount: { type: string, description: positive USDC amount, at most 6 decimals }
                fromAddress: { type: string, description: EVM address, 0x plus 40 hex }
                toAddress: { type: string, description: Stellar account id, G plus 55 base32 }
      responses:
        '200': { description: rawTx }
        '400': { description: body failed validation }
        '502': { description: Allbridge call failed }
components:
  securitySchemes:
    lobsterToken:
      type: apiKey
      in: header
      name: x-lobster-token
      description: >
        Shared token (LOBSTER_API_TOKEN). Also accepted as a bearer token or, for
        the SSE stream, a ?token= query parameter. Omitted when no token is set.
    bearerAuth:
      type: http
      scheme: bearer